hook customizado só aciona no android, backhandler
